refactored jwt-claims, added test to ci.yml, checked functions for exportability
refactored jwt-claims, added test to ci.yml

checked functions' exportability
