+++ .github/workflows/ci.yml
@@ -9,16 +9,25 @@ jobs:
       - uses: actions/checkout@v4
       - uses: actions/setup-go@v5
         with:
-          go-version: '1.24.0'
+          go-version: '1.24.1'
       - uses: golangci/golangci-lint-action@v4
         with:
           version: v1.64.5
           args: --timeout=5m
+  
+  test:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-go@v5
+        with:
+          go-version: '1.24.1'
+      - run: go test -race ./...
 
   build:
     strategy:
       matrix:
-        go-version: [ ^1.24.0 ]
+        go-version: [ ^1.24.1 ]
         os: [ ubuntu-latest ]
     runs-on: ${{ matrix.os }}
     steps:
